# openapi.yaml
openapi: 3.0.3
info:
  title: Zarza-Ski API
  version: 1.0.0
  description: >
    API for low-cost ski stay booking. Prices are calculated by the frontend:
    420 EUR per started week per person, 120 EUR per empty bed, child
    coefficients: age < 3 = 0.0, 3-8 = 0.8, age > 8 = 1.0.
servers:
  - url: http://localhost:8000/api
security:
  - bearerAuth: []
paths:
...
  /rooms/{room}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      summary: Show room
      responses:
        '200':
          $ref: '#/components/responses/RoomItem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  parameters:
...
    RoomId:
      name: room
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    RoomItem:
      description: Room resource
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/DataEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Room'
...
  schemas:
    DataEnvelope:
      type: object
      properties:
        data:
          type: object
...
    Room:
      type: object
      properties:
        id:
          type: integer
        num:
          type: integer
        nb_lits:
          type: integer
        building_id:
          type: integer
        floor:
          type: integer
        surface:
          type: integer
        view:
          type: string
          enum: [parking, mountains]
        balcony:
          type: boolean
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
    Image:
      type: object
      properties:
        id:
          type: integer
        room_id:
          type: integer
        name:
          type: string
        path:
          type: string
        url:
          type: string
          nullable: true
...